runtimes/python/requirements-sync.txt
falcon
pyyaml
uwsgi
jwt
jsonschema